requirements.txt
allure-robotframework==2.8.12
boto3==1.12.39
botocore==1.15.39
docutils==0.15.2
requests==2.23.0
robotframework==3.2.1
robotframework-datadriver==0.3.6
robotframework-selenium2library==3.0.0
robotframework-excellib==2.0.0
xlrd==1.2.0
hvac==0.10.3
kubernetes==11.0.0
pandas==1.3.2
psycopg2-binary==2.9.1
config-parser==0.0.1